requirements.txt
et_xmlfile==2.0.0
filelock==3.13.1
fsspec==2024.6.1
Jinja2==3.1.4
joblib==1.5.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
openpyxl==3.1.5
pandas==2.3.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
typing_extensions==4.12.2
tzdata==2025.2
